requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
fastapi==0.115.13
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.5.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.1
pyjwt==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-telegram-bot==22.1
pytz==2025.2
pyyaml==6.0.2
realtime==2.4.3
requests==2.32.4
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
strenum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.156
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.3
websockets==14.2
yarl==1.20.1
